Added a better error message if you try to use `lambdasync logs` without having a deployed function
