+++ bin/src/markdown/logs-with-no-function.md
@@ -1,0 +1,5 @@
+# function not deployed
+=====================================
+You need to deploy your function before you can get logs for it.
+
+Deploy your function by running `lambdasync` first.
